Corrected error in process(), lazy processing
There was an error here where the output channel data was being added to
both feedback filters, which means that data from the hipass filter was
"leaking" into the lowpass filter. Now each filter saves to a temporary
float variable to avoid crossover.

Also, if the valley size is set to the min, then processing is
completely disabled.