ValleySize now a float parameter, not frequency
At first glance, this change might seem a bit weird, because the valley
size is indeed a frequency, and the knob even has frequencies written on
it for the min/max value. However, the FrequencyParameter uses
logarithmic scaling, and since the minimum valley size is 0.1Hz (instead
of the usual 20Hz for filters), the first half of the knob rotation
represents a range of about 10Hz, which makes this very hard to control
in a performance setting.

In this case, using linear scaling for the frequency actually works much
better; most of the useful range of the parameter is in the first half,
but extreme cuts can still be reached in the second half.